requirements.txt
llama-index
llama-index-embeddings-huggingface
sentence-transformers
torch
flask
flask-cors
mistralai
python-dotenv
llama-index-llms-huggingface==0.2.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0
numpy>=2.0.0
